Iiboharz is fucking epic commit (Wiseau + Range Sources)
IIboharz has given me permission to share these assets with the mod, and as a result we now have Wiseau/Stockyard and Range in OFD.

I have wanted these maps in the mod for awhile anyways, and already gave permission to use the layout. But now that we can use this specific VMF we can easily finish this map for use in the mod in it's best, and prettiest state possible.

There are assets that will need to be replaced or reimplemented (in the case of live tf2 props predominantly) but the map is now available for all of us to work with and play on.